Upgrade for EF Core 6
